+++ packages/api/src/types.ts
@@ -3,7 +3,7 @@ import {
   AdxRecordValidatorDescription,
 } from '@adxp/schemas'
 import * as auth from '@adxp/auth'
-import { ErrorCode, GetRecordResponse } from './http-types.js'
+import { GetRecordResponse } from './http-types.js'
 
 export type SchemaOpt =
   | string
@@ -37,25 +37,4 @@ export interface BatchWrite {
   collection: string
   key?: string
   value?: any
-}
-
-export class WritePermissionError extends Error {
-  constructor() {
-    super('No write permissions have been granted for this repo')
-  }
-}
-
-export class APIResponseError extends Error {
-  constructor(
-    public httpStatusCode: number,
-    public httpStatusText: string,
-    public httpHeaders?: Record<string, string>,
-    public httpResponseBody?: any,
-  ) {
-    super(httpResponseBody?.message || `${httpStatusCode} ${httpStatusText}`)
-  }
-
-  get code(): ErrorCode {
-    return this.httpResponseBody?.code || 'NetworkError'
-  }
 }
